+++ changelog.md
@@ -1,8 +1,12 @@
-## 0.1.1
-### Changed:
+### 0.1.1
+#### Added
+- ``allequal`` and ``ndistinct`` constraints
+- ``except_0`` argument to ``alldifferent`` constraint
+
+#### Changed:
 - project description in readme.md
 - link to the html doc
 
-# 0.1.0
+## 0.1.0
 
 - initial release